Back, Close and Settings Buttons Working Properly

+++ Virtual Reality Experience/Assets/Scripts/UI/VolumeMixerSlider.cs
@@ -27,6 +27,6 @@ public class VolumeMixerSlider : MonoBehaviour
     public void VolumeSet()
     {
         audio.SetFloat("AudioVol", Mathf.Log10(slid.value) * 20 );
-        voltext.text = slid.value * 100 + "%";
+        voltext.text = (int)slid.value * 100 + "%";
     }
 }
